backend-logi/requirements.txt
# Framework Flask et extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1

# Base de données
SQLAlchemy>=2.0.40
psycopg2-binary
Flask-Migrate==4.0.5
# ==2.9.9

# Configuration
python-dotenv==1.0.0

# Sécurité - Chiffrement
cryptography==41.0.7

# HTTP requests (pour les APIs externes)
requests==2.31.0

# Tâches asynchrones
celery==5.3.4
redis==5.0.1

# Upload fichiers
cloudinary==1.36.0

# Génération PDF
reportlab==4.0.8

# Génération Excel
openpyxl==3.1.2

# QR Codes
qrcode[pil]==7.4.2

# Monitoring
sentry-sdk[flask]==1.38.0

# images
standard-imghdr

# ==================== NOTIFICATIONS ====================
# Installer selon les providers utilisés

# SMS - Twilio (SMS + WhatsApp)
twilio==8.10.0

# SMS - Vonage (anciennement Nexmo)
# vonage==3.12.0

# SMS - Africa's Talking (Afrique)
africastalking==1.2.5

# Email - SendGrid
sendgrid==6.11.0

# Email - AWS SES
boto3==1.34.0

# Push Notifications - Firebase Cloud Messaging
firebase-admin==6.2.0

# Push Notifications - Web Push (VAPID)
pywebpush==1.14.0

# WebSocket temps réel
flask-socketio==5.3.6
python-socketio==5.10.0

# Production server
gunicorn==21.2.0
eventlet==0.35.1

# Testing
pytest==7.4.3
hypothesis==6.88.1

# Note: Pour SMTP (Gmail, Outlook, etc.), aucune dépendance supplémentaire n'est requise
# Note: Pour OneSignal, seul 'requests' est nécessaire (déjà inclus)
